# Apps/packages/backend/api-doc.json
{"openapi":"3.0.0","paths":{"/article":{"get":{"operationId":"ArticleController_findAll","parameters":[{"name":"category","required":false,"in":"query","description":"query by category","schema":{}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleResDTO"}}}}},"tags":["Article"]}},"/article/search":{"post":{"operationId":"ArticleController_search","parameters":[],"requestBody":{"required":true,"description":"搜索内容","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SearchReqDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleResDTO"}}}}},"tags":["Article"]}},"/article/{id}":{"get":{"operationId":"ArticleController_findDetailById","parameters":[{"name":"id","required":true,"in":"path","description":"文章id","schema":{}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArtilceDetailResDTO"}}}}},"tags":["Article"]}},"/article/edit":{"post":{"operationId":"ArticleController_editArticle","parameters":[],"requestBody":{"required":true,"description":"文章实体","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleEditOrCreateReqDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionResDTO"}}}}},"tags":["Article"]}},"/article/create":{"post":{"operationId":"ArticleController_createArticle","parameters":[],"requestBody":{"required":true,"description":"文章实体","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ArticleEditOrCreateReqDTO"}}}},"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionResDTO"}}}}},"tags":["Article"]}},"/article/delete":{"post":{"operationId":"ArticleController_deleteArticle","parameters":[{"name":"id","required":true,"in":"path","description":"文章id","schema":{}}],"responses":{"default":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecutionResDTO"}}}}},"tags":["Article"]}},"/user/register":{"post":{"operationId":"UserController_regsiterUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserRegsiterDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/user/login":{"post":{"operationId":"UserController_loginUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserLoginDTO"}}}},"responses":{"200":{"description":""}},"tags":["User"]}},"/user/info":{"get":{"operationId":"UserController_getUserInfo","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/user/getInfo/{userId}":{"get":{"operationId":"UserController_getUserInfoById","parameters":[{"name":"userId","required":true,"in":"path","description":"用户ID","schema":{}}],"responses":{"200":{"description":""}},"tags":["User"]}},"/user/follow":{"post":{"operationId":"UserController_followUser","parameters":[],"requestBody":{"required":true,"description":"关注用户DTO","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowUserDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}},"/user/unfollow":{"post":{"operationId":"UserController_unFollowUser","parameters":[],"requestBody":{"required":true,"description":"关注用户DTO","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FollowUserDTO"}}}},"responses":{"201":{"description":""}},"tags":["User"]}}},"info":{"title":"开发者知识社区服务端API文档","description":"Powerby 东北大学 信息管理与信息系统","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"UserEntity":{"type":"object","properties":{"id":{"type":"number","default":2,"description":"用户id"},"username":{"type":"string","default":"","description":"用户名"},"password":{"type":"string","default":"","description":"密码"},"avatar":{"type":"string","default":"","description":"头像"},"nickname":{"type":"string","default":"","description":"社区名"},"work":{"type":"number","default":"","description":"职业"},"selfDesc":{"type":"string","default":"","description":"个人描述"},"interestsJson":{"type":"string","default":"","description":"感兴趣领域"},"isAdmin":{"type":"number","default":"","description":"是否为管理员"},"articles":{"default":"","description":"我的文章","type":"array","items":{"$ref":"#/components/schemas/ArticleEntity"}},"likedArticles":{"default":"","description":"我喜欢的文章","type":"array","items":{"$ref":"#/components/schemas/ArticleEntity"}},"collectArticles":{"default":"","description":"我收藏的文章","type":"array","items":{"$ref":"#/components/schemas/ArticleEntity"}},"fans":{"description":"我的粉丝","type":"array","items":{"$ref":"#/components/schemas/UserEntity"}},"follows":{"description":"我的关注","type":"array","items":{"$ref":"#/components/schemas/UserEntity"}}},"required":["id","username","password","avatar","nickname","work","selfDesc","interestsJson","isAdmin","articles","likedArticles","collectArticles","fans","follows"]},"CommentEntity":{"type":"object","properties":{"id":{"type":"number","default":2,"description":"评论id"},"content":{"type":"string","default":2,"description":"评论内容"},"to":{"default":2,"description":"评论目标用户","allOf":[{"$ref":"#/components/schemas/UserEntity"}]},"from":{"default":2,"description":"评论来自用户","allOf":[{"$ref":"#/components/schemas/UserEntity"}]},"article":{"default":2,"description":"评论文章","allOf":[{"$ref":"#/components/schemas/ArticleEntity"}]}},"required":["id","content","to","from","article"]},"ArticleEntity":{"type":"object","properties":{"id":{"type":"number","default":2,"description":"文章id"},"title":{"type":"string","default":"React hooks的那些事","description":"文章标题"},"desc":{"type":"string","default":"React hooks是一个新特性。。。","description":"文章描述"},"content":{"type":"string","default":"","description":"文章内容"},"comments":{"description":"评论","type":"array","items":{"$ref":"#/components/schemas/CommentEntity"}},"category":{"type":"number","enum":[1,2,3,4,5,6,7],"default":1,"description":"文章分类"},"likedBy":{"description":"点赞的用户","type":"array","items":{"$ref":"#/components/schemas/UserEntity"}},"collectBy":{"description":"收藏的用户","type":"array","items":{"$ref":"#/components/schemas/UserEntity"}},"author":{"description":"作者","allOf":[{"$ref":"#/components/schemas/UserEntity"}]}},"required":["id","title","desc","content","comments","category","likedBy","collectBy","author"]},"ArticleResDTO":{"type":"object","properties":{"code":{"type":"number"},"data":{"type":"array","items":{"$ref":"#/components/schemas/ArticleEntity"}}},"required":["code","data"]},"SearchReqDTO":{"type":"object","properties":{"search":{"type":"string","description":"搜索内容"},"category":{"type":"number","description":"文章类型"}},"required":["search","category"]},"ArtilceDetailResDTO":{"type":"object","properties":{"code":{"type":"number"},"data":{"$ref":"#/components/schemas/ArticleEntity"}},"required":["code","data"]},"ArticleEditOrCreateReqDTO":{"type":"object","properties":{"id":{"type":"number"},"article":{"$ref":"#/components/schemas/ArticleEntity"}},"required":["id","article"]},"ExecutionResDTO":{"type":"object","properties":{"code":{"type":"number"},"success":{"type":"boolean"}},"required":["code","success"]},"UserRegsiterDTO":{"type":"object","properties":{"user":{"description":"用户信息","allOf":[{"$ref":"#/components/schemas/UserEntity"}]}},"required":["user"]},"UserLoginDTO":{"type":"object","properties":{"username":{"type":"string","description":"账号"},"password":{"type":"string","description":"密码"}},"required":["username","password"]},"FollowUserDTO":{"type":"object","properties":{"fansId":{"type":"number","description":"粉丝id"},"followId":{"type":"number","description":"被关注的人id"}},"required":["fansId","followId"]}}}}